New particle preset and move them into their own file.
